refactor(CarRepository.java): 자동차들에 대한 책임을 지는 클래스 생성
CarRepository는 경주에서 사용될 자동차들에 대한 리스트를 가지고
자동차들에 대한 책임을 수행한다.

BREAKING CHANGE: RacingGame 클래스에서 List<Car> 필드를 같는 대신
CarRepository 필드를 갖는 걸로 수정
as-is:
public class RacingGame {

    private List<Car> cars;
    private int repeats;

    public RacingGame() {
        cars = new ArrayList<>();
        repeats = 0;
    }

   ...

    private void createCar() {
        final List<String> carNameList = Input.inputValidNames();
        carNameList.forEach((carName) -> cars.add(new Car(carName)));
    }

    private void startRacing() {
        ...
        while(repeats-- > 0) {
            cars.forEach(Car::goForward);
            Output.printTurnResult(cars);
        }
        Output.printWinner(findWinner());
    }

    private List<Car> findWinner() {
        final int farthestPosition = findFarthestPosition();
        return getWinners(farthestPosition);
    }

    private int findFarthestPosition() {
        return cars.stream()
            .sorted(Comparator.comparing(Car::getPosition))
            .collect(Collectors.toList())
            .get(cars.size() - 1)
            .getPosition();
    }

    private List<Car> getWinners(final int farthestPosition) {
        return cars.stream()
                .filter((car) -> farthestPosition == car.getPosition())
                .collect(Collectors.toList());
    }
}

to-be:
public class RacingGame {

    private CarRepository carRepository;
    private int repeats;

    public RacingGame() {
        repeats = 0;
    }

    public void startGame() {
        inputValidData();
        startRacing();
    }

    private void inputValidData() {
        createCar();
        repeats = Input.inputValidRepeats();
    }

    private void createCar() {
        final String carNames = Input.inputValidNames();
        carRepository = new CarRepository(carNames);
    }

    private void startRacing() {
        Output.printResultMessage();
        while(repeats-- > 0) {
            carRepository.startThisTurn();
            Output.printTurnResult(carRepository.getThisTurnResult());
        }
        Output.printWinner(carRepository.getWinners());
    }

}

Input:
as-is:
public class Input {

    ...

    public static List<String> inputValidNames() {
        final String names = inputNames();
        return Arrays.stream(names.split(RacingGameMessage.CAR_NAME_DISTRIBUTOR))
            .collect(Collectors.toList());
    }

    ...
}

public class Input {

    ...

    public static List<String> inputValidNames() {
        return inputNames();
    }

    ...
}
